+++ RELEASENOTES.md
@@ -2,12 +2,20 @@
 
 Powershell module to connect to all MS365 services and install required packages.
 
-(c) 2020 by [matthias fleschuetz](https://github.com/blindzero)
+(c) by [Matthias Fleschuetz](https://github.com/blindzero)
 [https://github.com/blindzero/Connect-MS365](https://github.com/blindzero/Connect-MS365)
 
 ## ToDos / Wishlist
 
 - Azure CLI
+
+## v1.2.0
+
+### New Features
+
+### Changes
+
+### Fixes
 
 ## v1.1.0